# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Technician Hub API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/technicians": {
      "get": {
        "tags": [
          "technicians"
        ],
        "summary": "List Technicians",
        "operationId": "list_technicians_api_technicians_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "pattern": "^[^\\x00-\\x1f\\x7f]*$",
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TechnicianSummary"
                  },
                  "title": "Response List Technicians Api Technicians Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "technicians"
        ],
        "summary": "Create Technician",
        "operationId": "create_technician_api_technicians_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicianCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicianDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technicians/{technician_id}": {
      "get": {
        "tags": [
          "technicians"
        ],
        "summary": "Get Technician",
        "operationId": "get_technician_api_technicians__technician_id__get",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicianDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "technicians"
        ],
        "summary": "Update Technician",
        "operationId": "update_technician_api_technicians__technician_id__patch",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicianUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicianDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "technicians"
        ],
        "summary": "Delete Technician",
        "operationId": "delete_technician_api_technicians__technician_id__delete",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicianDelete"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technicians/{technician_id}/calendar-assignments": {
      "get": {
        "tags": [
          "technicians"
        ],
        "summary": "Assignment History",
        "operationId": "assignment_history_api_technicians__technician_id__calendar_assignments_get",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentRead"
                  },
                  "title": "Response Assignment History Api Technicians  Technician Id  Calendar Assignments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technicians/{technician_id}/calendar": {
      "put": {
        "tags": [
          "technicians"
        ],
        "summary": "Set Calendar",
        "operationId": "set_calendar_api_technicians__technician_id__calendar_put",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "technicians"
        ],
        "summary": "Remove Calendar Assignment",
        "operationId": "remove_calendar_assignment_api_technicians__technician_id__calendar_delete",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/technician-forms/expense/submit": {
      "post": {
        "tags": [
          "expenses"
        ],
        "summary": "Submit",
        "operationId": "submit_api_technician_forms_expense_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technicians/{technician_id}/expenses": {
      "get": {
        "tags": [
          "expenses"
        ],
        "summary": "Expenses",
        "operationId": "expenses_api_technicians__technician_id__expenses_get",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/{expense_id}": {
      "get": {
        "tags": [
          "expenses"
        ],
        "summary": "Expense",
        "operationId": "expense_api_expenses__expense_id__get",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technicians/{technician_id}/expense-sessions/{session_id}/revoke": {
      "post": {
        "tags": [
          "expenses"
        ],
        "summary": "Revoke",
        "operationId": "revoke_api_technicians__technician_id__expense_sessions__session_id__revoke_post",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentInput": {
        "properties": {
          "calendar_id": {
            "type": "string",
            "format": "uuid",
            "title": "Calendar Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "calendar_id"
        ],
        "title": "AssignmentInput"
      },
      "AssignmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "technician_id": {
            "type": "string",
            "format": "uuid",
            "title": "Technician Id"
          },
          "calendar_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Id"
          },
          "calendar_name": {
            "type": "string",
            "title": "Calendar Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "technician_id",
          "calendar_id",
          "calendar_name",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "AssignmentRead"
      },
...
      "CalendarSummary": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "LOCAL_DEMO",
              "GOOGLE"
            ],
            "title": "Source",
            "default": "LOCAL_DEMO"
          },
          "availability": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "title": "Availability",
            "default": "AVAILABLE"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CalendarSummary"
      },
...
      "ExpenseInput": {
        "properties": {
          "expense_type": {
            "type": "string",
            "maxLength": 100,
            "title": "Expense Type"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "note": {
            "type": "string",
            "maxLength": 4000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expense_type",
          "amount"
        ],
        "title": "ExpenseInput"
      },
...
      "ExpenseReceipt": {
        "properties": {
          "expense_id": {
            "type": "string",
            "format": "uuid",
            "title": "Expense Id"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number",
            "default": 1
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Expense saved successfully."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expense_id"
        ],
        "title": "ExpenseReceipt"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TechnicianCreate": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "calendar_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "first_name",
          "last_name"
        ],
        "title": "TechnicianCreate"
      },
      "TechnicianDelete": {
        "properties": {
          "confirmation": {
            "type": "string",
            "const": "DELETE",
            "title": "Confirmation"
          },
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirmation",
          "expected_updated_at"
        ],
        "title": "TechnicianDelete"
      },
      "TechnicianDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "title": "Status"
          },
          "calendar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalendarSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "integrations": {
            "$ref": "#/components/schemas/IntegrationSummary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "accounting_timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accounting Timezone"
          },
          "driver_license_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver License Id"
          },
          "ssn_last4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssn Last4"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "photo_url",
          "status",
          "calendar",
          "integrations",
          "created_at",
          "updated_at",
          "driver_license_id",
          "ssn_last4"
        ],
        "title": "TechnicianDetail"
      },
      "TechnicianSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "title": "Status"
          },
          "calendar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalendarSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "integrations": {
            "$ref": "#/components/schemas/IntegrationSummary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "photo_url",
          "status",
          "calendar",
          "integrations",
          "created_at",
          "updated_at"
        ],
        "title": "TechnicianSummary"
      },
      "TechnicianUpdate": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "INACTIVE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "driver_license_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver License Id"
          },
          "ssn_last4": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssn Last4"
          },
          "accounting_timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Accounting Timezone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TechnicianUpdate"
      },
...
    }
  }
}